requirements.txt
# /build.py
mpypack~=0.1.1

# /mgui/dev/screen_server.py
aiohttp~=3.7.1
# /mgui/dev/framebuf_helper.py
pillow~=8.0.1
# /mgui/dev/ssd1306pygame.py
pygame~=2.0.1